// src/pages/registrathion/types.d.ts
import Button from '../../components/button';
import InputFormProps from '../../components/input-form';
import { TagAttribute, ComponentEvent } from '../../core/types';

export type RegistrathionProps = {
  tagAttrs?: TagAttribute
  InputFormEmail: InputFormProps
  InputFormLogin: InputFormProps
  InputFormFirstName: InputFormProps
  InputFormSecondName : InputFormProps
  InputFormPhone: InputFormProps
  InputFormPassword: InputFormProps
  InputFormPasswordAgain: InputFormProps
  RegistrathionButton: Button
  LoginButton: Button
  events?: ComponentEvent
}